+++ apis/r/R/SOMADenseNDArray.R
@@ -1,4 +1,4 @@
-#' SOMADenseNDArray
+#' SOMA Dense Nd-Array
 #'
 #' @description \code{SOMADenseNDArray} is a dense, N-dimensional array of
 #' a \code{primitive} type, with offset (zero-based) \code{int64} integer
@@ -18,11 +18,8 @@
 #' more dimensions.
 #'
 #' The default \dQuote{fill} value for \code{SOMADenseNDArray} is the zero or
-#' null value of the array type (e.g. \code{\link[arrow]{float32}} defaults
-#' to 0.0).
-#'
-#' The \code{$write()} method is currently limited to writing from 2-d matrices
-#' (lifecycle: maturing).
+#' null value of the array type (e.g.,
+#' \code{\link[arrow:float32]{arrow::float32}()} defaults to 0.0).
 #'
 #' @param coords Optional \code{list} of integer vectors, one for each
 #' dimension, with a length equal to the number of values to read. If
@@ -51,7 +48,7 @@ SOMADenseNDArray <- R6::R6Class(
       result_order = "auto",
       log_level = "auto"
     ) {
-      private$check_open_for_read()
+      private$.check_open_for_read()
 
       uri <- self$uri
 
@@ -92,14 +89,17 @@ SOMADenseNDArray <- R6::R6Class(
       result_order = "ROW_MAJOR",
       log_level = "warn"
     ) {
-      private$check_open_for_read()
+      private$.check_open_for_read()
 
       ndim <- self$ndim()
       attrnames <- self$attrnames()
 
       stopifnot(
         "Array must have two dimensions" = ndim == 2,
-        "Array must contain column 'soma_data'" = all.equal("soma_data", attrnames)
+        "Array must contain column 'soma_data'" = all.equal(
+          "soma_data",
+          attrnames
+        )
       )
 
       if (is.null(coords)) {
@@ -112,28 +112,37 @@ SOMADenseNDArray <- R6::R6Class(
         ncol <- length(unique(as.numeric(coords[[2]])))
       }
 
-      tbl <- self$read_arrow_table(coords = coords, result_order = result_order, log_level = log_level)
-      m <- matrix(as.numeric(tbl$GetColumnByName("soma_data")),
-        nrow = nrow, ncol = ncol,
-        byrow = result_order == "ROW_MAJOR"
+      tbl <- self$read_arrow_table(
+        coords = coords,
+        result_order = result_order,
+        log_level = log_level
       )
+      return(matrix(
+        as.numeric(tbl$GetColumnByName("soma_data")),
+        nrow = nrow,
+        ncol = ncol,
+        byrow = result_order == "ROW_MAJOR"
+      ))
     },
 
-    #' @description Write matrix data to the array (lifecycle: maturing).
+    #' @description Write matrix data to the array (lifecycle: maturing).\cr
+    #' \cr
+    #' \strong{Note}: The \code{$write()} method is currently limited to writing
+    #' from two-dimensional matrices (lifecycle: maturing).
     #'
     # More general write methods for higher-dimensional array could be added.
     #'
     #' @param values A \code{matrix}. Character dimension names are ignored
     #' because \code{SOMANDArray}s use integer indexing.
     #' @param coords A \code{list} of integer vectors, one for each dimension,
-    #' with a length equal to the number of values to write. If \code{NULL}, the
-    #' default, the values are taken from the row and column names of
-    #' \code{values}.
+    #' with a length equal to the number of values to write. If \code{NULL},
+    #' the default, the values are taken from the row and column names
+    #' of \code{values}.
     #'
-    #' @return Invisibly returns \code{self}
+    #' @return Invisibly returns \code{self}.
     #'
     write = function(values, coords = NULL) {
-      private$check_open_for_write()
+      private$.check_open_for_write()
 
       spdl::debug("[SOMADenseNDArray::write] entered")
       stopifnot(
@@ -156,7 +165,7 @@ SOMADenseNDArray <- R6::R6Class(
         private$.type <- self$schema()[["soma_data"]]$type
       }
 
-      arr <- self$object
+      arr <- private$.tiledb_array
       tiledb::query_layout(arr) <- "COL_MAJOR"
       spdl::debug("[SOMADenseNDArray::write] about to call write")
       arrsch <- arrow::schema(arrow::field("soma_data", private$.type))
@@ -166,7 +175,6 @@ SOMADenseNDArray <- R6::R6Class(
       naap <- nanoarrow::nanoarrow_allocate_array()
       nasp <- nanoarrow::nanoarrow_allocate_schema()
       arrow::as_record_batch(tbl)$export_to_c(naap, nasp)
-      # arr[] <- values
       writeArrayFromArrow(
         uri = self$uri,
         naap = naap,
@@ -178,15 +186,10 @@ SOMADenseNDArray <- R6::R6Class(
       )
       spdl::debug("[SOMADenseNDArray::write] written")
 
-      # tiledb-r always closes the array after a write operation so we need to
-      # manually reopen it until close-on-write is optional
-      # self$open("WRITE", internal_use_only = "allowed_use")
-      invisible(self)
+      return(invisible(self))
     }
   ),
   private = list(
-    .is_sparse = FALSE,
-
     # Given a user-specified shape along a particular dimension, returns a named
     # list containing name, capacity, and extent elements. The shape cannot be
     # NULL for dense arrays.